+++ src/main.rs
@@ -1,48 +1,35 @@
 use axum::body::{Body, Bytes};
-use axum::extract::{Query, Request};
+use axum::extract::Request;
 use axum::http::HeaderMap;
 use axum::middleware::Next;
 use axum::routing::get;
-use axum::Json;
 use axum::{
-    extract::{Path, State},
     http::StatusCode,
     response::{IntoResponse, Response},
-    routing::post,
     Router,
 };
 use clap::Parser;
 
 use error::AppError;
-use forward::message::ReforwardInfo;
-use http::Uri;
 use http_body_util::BodyExt;
 use local_ip_address::local_ip;
-use std::collections::HashMap;
 use std::env;
 use std::future::IntoFuture;
 use std::net::SocketAddr;
 use std::str::FromStr;
 use std::sync::Arc;
-use tracing_subscriber::EnvFilter;
-
-#[cfg(debug_assertions)]
-use tower_http::services::{ServeDir, ServeFile};
-
 use tower_http::cors::CorsLayer;
 use tower_http::trace::TraceLayer;
 use tower_http::validate_request::ValidateRequestHeaderLayer;
 use tracing::info_span;
 use tracing::{debug, error, info};
-use webrtc::peer_connection::sdp::session_description::RTCSessionDescription;
+use tracing_subscriber::EnvFilter;
 
 use crate::auth::ManyValidate;
 use crate::config::Config;
-use crate::result::Result;
-use config::IceServer;
+use crate::route::r#static::static_server;
+use crate::route::{admin, resource, whep, whip, AppState};
 use stream::manager::Manager;
-#[cfg(not(debug_assertions))]
-use {http::header, rust_embed::RustEmbed};
 
 mod auth;
 mod config;
@@ -53,6 +40,7 @@ mod forward;
 mod hook;
 mod metrics;
 mod result;
+mod route;
 mod stream;
 
 #[derive(Parser)]
@@ -92,27 +80,13 @@ async fn main() {
     let admin_auth_layer =
         ValidateRequestHeaderLayer::custom(ManyValidate::new(vec![cfg.admin_auth]));
     let app = Router::new()
-        .route(&live777_http::path::whip(":stream"), post(whip))
-        .route(&live777_http::path::whep(":stream"), post(whep))
-        .route(
-            &live777_http::path::resource(":stream", ":session"),
-            post(change_resource)
-                .patch(add_ice_candidate)
-                .delete(remove_stream_session),
+        .merge(
+            whip::route()
+                .merge(whep::route())
+                .merge(resource::route())
+                .layer(auth_layer),
         )
-        .route(
-            &live777_http::path::resource_layer(":stream", ":session"),
-            get(get_layer).post(select_layer).delete(un_select_layer),
-        )
-        .layer(auth_layer)
-        .route(
-            live777_http::path::ADMIN_INFOS,
-            get(infos).layer(admin_auth_layer.clone()),
-        )
-        .route(
-            &live777_http::path::reforward(":stream"),
-            post(reforward).layer(admin_auth_layer),
-        )
+        .merge(admin::route().layer(admin_auth_layer))
         .route(live777_http::path::METRICS, get(metrics))
         .with_state(app_state.clone())
         .layer(if cfg.http.cors {
@@ -177,45 +151,6 @@ async fn metrics() -> String {
         .unwrap()
 }
 
-#[cfg(not(debug_assertions))]
-#[derive(RustEmbed)]
-#[folder = "assets/"]
-struct Assets;
-
-fn static_server(router: Router) -> Router {
-    #[cfg(debug_assertions)]
-    {
-        let serve_dir =
-            ServeDir::new("assets").not_found_service(ServeFile::new("assets/index.html"));
-        router.nest_service("/", serve_dir.clone())
-    }
-    #[cfg(not(debug_assertions))]
-    {
-        router.fallback(static_handler)
-    }
-}
-
-#[cfg(not(debug_assertions))]
-async fn static_handler(uri: Uri) -> impl IntoResponse {
-    let mut path = uri.path().trim_start_matches('/');
-    if path.is_empty() {
-        path = "index.html";
-    }
-    match Assets::get(path) {
-        Some(content) => {
-            let mime = mime_guess::from_path(path).first_or_octet_stream();
-            ([(header::CONTENT_TYPE, mime.as_ref())], content.data).into_response()
-        }
-        None => (StatusCode::NOT_FOUND, "not found").into_response(),
-    }
-}
-
-#[derive(Clone)]
-struct AppState {
-    config: Config,
-    stream_manager: Arc<Manager>,
-}
-
 async fn print_request_response(
     req: Request,
     next: Next,
@@ -258,230 +193,4 @@ where
     }
 
     Ok(bytes)
-}
-
-async fn whip(
-    State(state): State<AppState>,
-    Path(stream): Path<String>,
-    header: HeaderMap,
-    body: String,
-) -> Result<Response<String>> {
-    let content_type = header
-        .get("Content-Type")
-        .ok_or(anyhow::anyhow!("Content-Type is required"))?;
-    if content_type.to_str()? != "application/sdp" {
-        return Err(anyhow::anyhow!("Content-Type must be application/sdp").into());
-    }
-    let offer = RTCSessionDescription::offer(body)?;
-    let (answer, session) = state.stream_manager.publish(stream.clone(), offer).await?;
-    let mut builder = Response::builder()
-        .status(StatusCode::CREATED)
-        .header("Content-Type", "application/sdp")
-        .header("Accept-Patch", "application/trickle-ice-sdpfrag")
-        .header("Location", live777_http::path::resource(&stream, &session));
-    for link in link_header(state.config.ice_servers.clone()) {
-        builder = builder.header("Link", link);
-    }
-    Ok(builder.body(answer.sdp)?)
-}
-
-async fn whep(
-    State(state): State<AppState>,
-    Path(stream): Path<String>,
-    header: HeaderMap,
-    body: String,
-) -> Result<Response<String>> {
-    let content_type = header
-        .get("Content-Type")
-        .ok_or(anyhow::anyhow!("Content-Type is required"))?;
-    if content_type.to_str()? != "application/sdp" {
-        return Err(anyhow::anyhow!("Content-Type must be application/sdp").into());
-    }
-    let offer = RTCSessionDescription::offer(body)?;
-    let (answer, session) = state
-        .stream_manager
-        .subscribe(stream.clone(), offer)
-        .await?;
-    let mut builder = Response::builder()
-        .status(StatusCode::CREATED)
-        .header("Content-Type", "application/sdp")
-        .header("Accept-Patch", "application/trickle-ice-sdpfrag")
-        .header("Location", live777_http::path::resource(&stream, &session));
-    for link in link_header(state.config.ice_servers.clone()) {
-        builder = builder.header("Link", link);
-    }
-    if state.stream_manager.layers(stream.clone()).await.is_ok() {
-        builder = builder.header(
-            "Link",
-            format!(
-                "<{}>; rel=\"urn:ietf:params:whep:ext:core:layer\"",
-                live777_http::path::resource_layer(&stream, &session)
-            ),
-        )
-    }
-    Ok(builder.body(answer.sdp)?)
-}
-
-async fn add_ice_candidate(
-    State(state): State<AppState>,
-    Path((stream, session)): Path<(String, String)>,
-    header: HeaderMap,
-    body: String,
-) -> Result<Response<String>> {
-    let content_type = header
-        .get("Content-Type")
-        .ok_or(AppError::from(anyhow::anyhow!("Content-Type is required")))?;
-    if content_type.to_str()? != "application/trickle-ice-sdpfrag" {
-        return Err(anyhow::anyhow!("Content-Type must be application/trickle-ice-sdpfrag").into());
-    }
-    state
-        .stream_manager
-        .add_ice_candidate(stream, session, body)
-        .await?;
-    Ok(Response::builder()
-        .status(StatusCode::NO_CONTENT)
-        .body("".to_string())?)
-}
-
-async fn remove_stream_session(
-    State(state): State<AppState>,
-    Path((stream, session)): Path<(String, String)>,
-    _uri: Uri,
-) -> Result<Response<String>> {
-    state
-        .stream_manager
-        .remove_stream_session(stream, session)
-        .await?;
-    Ok(Response::builder()
-        .status(StatusCode::NO_CONTENT)
-        .body("".to_string())?)
-}
-
-async fn change_resource(
-    State(state): State<AppState>,
-    Path((stream, session)): Path<(String, String)>,
-    Json(req): Json<live777_http::request::ChangeResource>,
-) -> Result<Json<HashMap<String, String>>> {
-    state
-        .stream_manager
-        .change_resource(stream, session, (req.kind, req.enabled))
-        .await?;
-    Ok(Json(HashMap::new()))
-}
-
-async fn get_layer(
-    State(state): State<AppState>,
-    Path((stream, _session)): Path<(String, String)>,
-) -> Result<Json<Vec<live777_http::response::Layer>>> {
-    Ok(Json(
-        state
-            .stream_manager
-            .layers(stream)
-            .await?
-            .into_iter()
-            .map(|layer| layer.into())
-            .collect(),
-    ))
-}
-
-async fn select_layer(
-    State(state): State<AppState>,
-    Path((stream, session)): Path<(String, String)>,
-    Json(req): Json<live777_http::request::SelectLayer>,
-) -> Result<String> {
-    state
-        .stream_manager
-        .select_layer(
-            stream,
-            session,
-            req.encoding_id
-                .map(|encoding_id| forward::message::Layer { encoding_id }),
-        )
-        .await?;
-    Ok("".to_string())
-}
-
-async fn un_select_layer(
-    State(state): State<AppState>,
-    Path((stream, session)): Path<(String, String)>,
-) -> Result<String> {
-    state
-        .stream_manager
-        .select_layer(
-            stream,
-            session,
-            Some(forward::message::Layer {
-                encoding_id: constant::RID_DISABLE.to_string(),
-            }),
-        )
-        .await?;
-    Ok("".to_string())
-}
-
-async fn infos(
-    State(state): State<AppState>,
-    Query(req): Query<live777_http::request::QueryInfo>,
-) -> Result<Json<Vec<live777_http::response::StreamInfo>>> {
-    Ok(Json(
-        state
-            .stream_manager
-            .info(req.streams.map_or(vec![], |streams| {
-                streams
-                    .split(',')
-                    .map(|stream| stream.to_string())
-                    .collect()
-            }))
-            .await
-            .into_iter()
-            .map(|forward_info| forward_info.into())
-            .collect(),
-    ))
-}
-
-async fn reforward(
-    State(state): State<AppState>,
-    Path(stream): Path<String>,
-    Json(req): Json<live777_http::request::Reforward>,
-) -> Result<String> {
-    state
-        .stream_manager
-        .reforward(
-            stream,
-            ReforwardInfo {
-                target_url: req.target_url,
-                admin_authorization: req.admin_authorization,
-                resource_url: None,
-            },
-        )
-        .await?;
-    Ok("".to_string())
-}
-
-fn link_header(ice_servers: Vec<IceServer>) -> Vec<String> {
-    ice_servers
-        .into_iter()
-        .flat_map(|server| {
-            let mut username = server.username;
-            let mut credential = server.credential;
-            if !username.is_empty() {
-                username = string_encoder(&username);
-                credential = string_encoder(&credential);
-            }
-            server.urls.into_iter().map(move |url| {
-                let mut link = format!("<{}>; rel=\"ice-server\"", url);
-                if !username.is_empty() {
-                    link = format!(
-                        "{}; username=\"{}\"; credential=\"{}\"; credential-type=\"{}\"",
-                        link, username, credential, server.credential_type
-                    );
-                }
-                link
-            })
-        })
-        .collect()
-}
-
-fn string_encoder(s: &impl ToString) -> String {
-    let s = serde_json::to_string(&s.to_string()).unwrap();
-    s[1..s.len() - 1].to_string()
 }
