+++ src/route/mod.rs
@@ -1,0 +1,44 @@
+use crate::config::{Config, IceServer};
+use crate::stream::manager::Manager;
+use std::sync::Arc;
+
+pub mod admin;
+pub mod resource;
+pub mod r#static;
+pub mod whep;
+pub mod whip;
+
+#[derive(Clone)]
+pub struct AppState {
+    pub config: Config,
+    pub stream_manager: Arc<Manager>,
+}
+
+fn link_header(ice_servers: Vec<IceServer>) -> Vec<String> {
+    ice_servers
+        .into_iter()
+        .flat_map(|server| {
+            let mut username = server.username;
+            let mut credential = server.credential;
+            if !username.is_empty() {
+                username = string_encoder(&username);
+                credential = string_encoder(&credential);
+            }
+            server.urls.into_iter().map(move |url| {
+                let mut link = format!("<{}>; rel=\"ice-server\"", url);
+                if !username.is_empty() {
+                    link = format!(
+                        "{}; username=\"{}\"; credential=\"{}\"; credential-type=\"{}\"",
+                        link, username, credential, server.credential_type
+                    );
+                }
+                link
+            })
+        })
+        .collect()
+}
+
+fn string_encoder(s: &impl ToString) -> String {
+    let s = serde_json::to_string(&s.to_string()).unwrap();
+    s[1..s.len() - 1].to_string()
+}
